+++ assets/GAMETEXT/base/dinomod.patch.spec
